Fix posix_lseek result overflow
Seen when testing Spider-Man Miles Morales on more-kernel.

+++ src/core/libraries/kernel/file_system.cpp
@@ -237,7 +237,7 @@ s64 PS4_SYSV_ABI sceKernelLseek(int d, s64 offset, int whence) {
 }
 
 s64 PS4_SYSV_ABI posix_lseek(int d, s64 offset, int whence) {
-    int result = sceKernelLseek(d, offset, whence);
+    s64 result = sceKernelLseek(d, offset, whence);
     if (result < 0) {
         LOG_ERROR(Kernel_Pthread, "posix_lseek: error = {}", result);
         ErrSceToPosix(result);
